+++ packages/designer/tests/utils/renderer.ts
@@ -2,7 +2,7 @@ export function getMockRenderer() {
   return {
     isSimulatorRenderer: true,
     run() {
-      console.log('renderer run');
+      // console.log('renderer run');
     }
   }
 }